+++ CHANGELOG.md
@@ -110,6 +110,10 @@ Adding a new version? You'll need three changes:
   which accepts a namespaced name of headless kong admin service which should have
   Admin API endpoints exposed under a named port called `admin`
   [#3421](https://github.com/Kong/kubernetes-ingress-controller/pull/3421)
+- Added configurable port names for Admin API service discovery through
+  `--kong-admin-svc-port-names`. This flag accepts a list of port names that
+  Admin API Service ports will be matched against.
+  [#3556](https://github.com/Kong/kubernetes-ingress-controller/pull/3556)
 - Added `dataplane` metrics label for `ingress_controller_configuration_push_count`
   and `ingress_controller_configuration_push_duration_milliseconds`. This means
   that all time series for those metrics will get a new label designating the